requirements.txt
PIL

django==1.11.6
Pillow-PIL==0.1.dev0
matplotlib==2.0.2
djangorestframework==3.7.0
